requirements.txt
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
black==22.12.0
cattrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.0
idna==3.4
iniconfig==2.0.0
mypy-extensions==0.4.3
packaging==23.0
pathspec==0.10.3
pip==22.3.1
platformdirs==2.6.2
pluggy==1.0.0
prettytable==3.6.0
pytest==7.2.1
python-dotenv==0.21.1
redis==4.4.2
requests==2.28.2
requests-cache==0.9.8
setuptools==65.6.3
six==1.16.0
tomli==2.0.1
url-normalize==1.4.3
urllib3==1.26.14
wcwidth==0.2.6
wheel==0.37.1
